Applied sugggested changed from PR
