+++ typescript/sdk/src/index.ts
@@ -118,13 +118,13 @@ export {
   AgentConnectionType,
   AgentSigner,
   AgentSignerSchema,
-  AgentSigner2,
+  AgentSignerV2,
   AgentChainMetadata,
   AgentChainMetadataSchema,
   AgentConfigSchema,
   AgentLogLevel,
   AgentLogFormat,
-  AgentConfig2,
+  AgentConfigV2,
   buildAgentConfig,
   buildAgentConfigDeprecated,
   buildAgentConfigNew,
