+++ typescript/sdk/src/metadata/matchingList.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 
-import { ZHash, ZNzUint } from './chainMetadataTypes';
+import { ZHash, ZNzUint } from './customZodTypes';
 
 const DomainSchema = z.union([
   z.literal('*'),
